Migrate Not query building to new interface
